@@ -2,45 +2,40 @@
   "name": "@zlux/widgets",
   "version": "0.0.4",
   "scripts": {
-    "start": "webpack --progress --colors --watch",
-    "build": "webpack --progress --colors",
+    "start": "webpack --progress --color --watch",
+    "build": "webpack --progress --color",
     "metadata": "ngc -p tsconfig.ngx-i18n.json",
-    "lint": "ng lint"
+    "lint": "tslint -p ./tsconfig.json -c ./tslint.json"
   },
   "license": "EPL-2.0",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "dependencies": {
-    "@angular/animations": "~6.0.9",
-    "@angular/common": "~6.0.9",
-    "@angular/core": "~6.0.9",
-    "@angular/forms": "~6.0.9",
-    "@angular/platform-browser": "~6.0.9",
-    "@angular/router": "~6.0.9",
-    "rxjs": "~6.2.2",
-    "rxjs-compat": "~6.2.2",
-    "zone.js": "~0.8.26"
+    "@angular/animations": "~12.0.0",
+    "@angular/common": "~12.0.0",
+    "@angular/core": "~12.0.0",
+    "@angular/forms": "~12.0.0",
+    "@angular/platform-browser": "~12.0.0",
+    "@angular/router": "~12.0.0",
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular/compiler": "~6.0.9",
-    "@angular/compiler-cli": "~6.0.9",
-    "@types/node": "~6.0.60",
+    "@angular/compiler": "~12.0.0",
+    "@angular/compiler-cli": "~12.0.0",
+    "@types/node": "~12.0.0",
+    "ajv": "~8.8.2",
     "angular2-template-loader": "~0.6.2",
-    "codelyzer": "~4.4.2",
-    "copy-webpack-plugin": "~4.5.2",
-    "css-loader": "~1.0.0",
-    "css-to-string-loader": "~0.1.3",
-    "exports-loader": "~0.7.0",
-    "extract-text-webpack-plugin": "~3.0.2",
-    "file-loader": "~1.1.11",
-    "html-loader": "~0.5.5",
-    "html-webpack-plugin": "~3.2.0",
-    "style-loader": "~0.21.0",
-    "svg-sprite-loader": "~3.8.0",
-    "ts-loader": "~3.5.0",
-    "ts-node": "~3.0.4",
+    "codelyzer": "~6.0.2",
+    "copy-webpack-plugin": "~10.0.0",
+    "css-loader": "~6.5.1",
+    "html-loader": "~3.0.1",
+    "rxjs": "~6.6.0",
+    "style-loader": "~3.3.1",
+    "ts-loader": "~8.3.0",
     "tslint": "~5.10.0",
-    "typescript": "~2.7.2",
-    "url-loader": "~1.0.1",
-    "webpack": "~3.8.0",
-    "webpack-shell-plugin": "~0.5.0"
+    "typescript": "~4.2.0",
+    "webpack": "~5.64.4",
+    "webpack-cli": "~4.9.1",
+    "webpack-shell-plugin-next": "~2.2.2"
   }
 }
