@@ -10,19 +10,25 @@
   Copyright Contributors to the Zowe Project.
 */
 
-import { NgModule, Directive, Input } from "@angular/core";
+import { NgModule, Directive, Input, ElementRef, OnInit } from "@angular/core";
 import { CommonModule } from '@angular/common';
-import '../../assets/css/inputtext.css';
+import styles from '../../assets/css/inputtext.module.css';
 
 @Directive({
-  selector: '[zluxInputText]',
-  host: {
-    '[class.text-input]': 'true',
-    '[class.text-input-dark]': 'dark'
-  }
+  selector: '[zluxInputText]'
 })
-export class ZluxInputText {
+export class ZluxInputText implements OnInit {
   @Input() dark: boolean = false;
+
+  constructor(private element: ElementRef) {
+    this.element.nativeElement.className += ' ' + styles['text-input'];
+  }
+
+  ngOnInit(): void {
+    if (this.dark) {
+      this.element.nativeElement.className += ' ' + styles['text-input-dark'];
+    }
+  }
 }
 
 @NgModule({
