@@ -10,11 +10,11 @@
   Copyright Contributors to the Zowe Project.
 */
 
-var webpack = require('webpack');
-var path = require('path');
-const ExtractTextPlugin = require('extract-text-webpack-plugin');
+const webpack = require('webpack');
+const path = require('path');
 const CopyWebpackPlugin = require('copy-webpack-plugin');
-const WebpackShellPlugin = require('webpack-shell-plugin');
+const WebpackShellPlugin = require('webpack-shell-plugin-next');
+
 
 function root(__path) {
   return path.join(__dirname, __path);
@@ -23,9 +23,10 @@ function root(__path) {
 var config = {
   entry: [root('./src/app/index.ts')],
   devtool: 'source-map',
+  mode: 'production',
   output: {
     path: root('.'),
-    filename: 'index.js',
+    filename: './dist/index.js',
     library: '@zlux/widgets',
     libraryTarget: 'amd',
     umdNamedDefine: true
@@ -34,69 +35,83 @@ var config = {
     '@angular/core',
     '@angular/common',
     '@angular/compiler',
-    '@angular/forms'
+    '@angular/forms',
+    'rxjs',
+    'rxjs/operators'
   ],
   module: {
     rules: [
       {
         test: /\.ts$/,
-        loaders: [
+        exclude: /node_modules/,
+        use: [
           'ts-loader',
           'angular2-template-loader'
         ]
       },
       {
         test: /\.html$/,
-        loader: 'html-loader'
+        use: [{
+          loader: 'html-loader',
+          options: {
+            esModule: false
+          }
+        }]
       },
       {
         test: /\.svg$/,
-        loader: 'svg-sprite-loader'
+        type: 'asset/inline'
       },
       {
         /* External (or inline) file loader */
-        'test': /\.(jpg|png|gif|otf|ttf|woff|woff2|cur|ani)$/,
-        'loader': 'url-loader?name=[name].[hash:20].[ext]&limit=10000',
-        'options':{
-          useRelativePath: true
-        }
-      },
-      {
-        test: /\.css$/,
-        exclude: root('./src/app'),
-        loader: ExtractTextPlugin.extract({ use: 'css-loader?sourceMap', fallback: 'style-loader' })
+        test: /\.(jpg|png|gif|otf|ttf|woff|woff2|cur|ani)$/,
+        type: 'asset',
       },
       {
         test: /\.css$/,
         include: root('./src/app/assets/css'),
-        loader: 'style-loader!css-loader'
+        use: [
+          'style-loader',
+          {
+            loader: 'css-loader',
+            options: {
+              modules: true,
+              sourceMap: false
+            }
+          }
+        ]
       },
       {
         test: /\.css$/,
         include: root('./src/app/components'),
-        loader: 'css-to-string-loader!css-loader'
+        use: [{
+          loader: 'css-loader',
+          options: {
+            exportType: 'string',
+            esModule: false,
+            sourceMap: false
+          }
+        }]
       }
     ]
   },
   resolve: {
     modules: [
-      "node_modules"
+      'node_modules'
     ],
-    extensions: [".js", ".ts"]
+    extensions: ['.js', '.ts']
   },
   plugins: [
-    new ExtractTextPlugin("styles.css"),
-    new webpack.WatchIgnorePlugin([
-      /\.js$/,
-      /\.d\.ts$/
-    ]),
-    new CopyWebpackPlugin([
-      {
+    new webpack.WatchIgnorePlugin({
+      paths: [/\.js$/, /\.d\.ts$/]
+    }),
+    new CopyWebpackPlugin({
+      patterns: [{
         from: '**/*.metadata.json',
-        to: '.',
+        to: './dist',
         context: './out_tsc/src/app/'
-      }
-    ]),
+      }]
+    }),
     new WebpackShellPlugin({
       onBuildStart: ['npm run metadata']
     })
@@ -110,9 +125,9 @@ module.exports = config;
   This program and the accompanying materials are
   made available under the terms of the Eclipse Public License v2.0 which accompanies
   this distribution, and is available at https://www.eclipse.org/legal/epl-v20.html
-  
+
   SPDX-License-Identifier: EPL-2.0
-  
+
   Copyright Contributors to the Zowe Project.
 */
 
